Clean up unused files and dependencies after search refactor
