--- functions/api/_search-utils.js
@@ -1,11 +1,0 @@
-/**
- * Search Utilities (Facade)
- * Re-exports from modular utility files for backward compatibility.
- * @version 7.0.0
- */
-
-export * from './_search-query.js';
-export * from './_search-content.js';
-export * from './_search-url.js';
-export * from './_search-scoring.js';
-export * from './_search-results.js';
